agent/requirements.txt
# AI Agent 项目依赖

# 核心框架
langchain==0.1.0
openai==1.3.0
chromadb==0.4.18

# Web框架
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# 数据处理
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0

# 工具和库
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
python-dotenv==1.0.0

# 向量数据库
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 日志和监控
loguru==0.7.2

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1 